WIP: Remove bottom banner view from current view stack
